# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: IAERP API
  version: 0.1.0-draft
  description: Contrato preliminar del MVP. El tenant activo se deriva del token.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.example.invalid/api/v1
security:
  - oauth2: []
paths:
  /context:
    get:
      operationId: getContext
      summary: Obtener tenant, permisos, limites y kill switch
      security:
        - oauth2: ["context:read"]
      responses:
        "200":
          description: Contexto activo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantContext"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /memberships:
    get:
      operationId: listOwnMemberships
      summary: Listar membresias activas del usuario
      security:
        - oauth2: ["memberships:read"]
      responses:
        "200":
          description: Membresias visibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Membership"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /service-accounts:
    get:
      operationId: listServiceAccounts
      summary: Listar service accounts del tenant activo
      security:
        - oauth2: ["service-accounts:read"]
      responses:
        "200":
          description: Service accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createServiceAccount
      summary: Crear service account ligado al tenant activo
      security:
        - oauth2: ["service-accounts:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccountInput"
      responses:
        "201":
          description: Service account creado; el secreto solo se entrega una vez
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /service-accounts/{serviceAccountId}:
    delete:
      operationId: revokeServiceAccount
      summary: Revocar un service account del tenant activo
      security:
        - oauth2: ["service-accounts:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: serviceAccountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Service account revocado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /establishments:
    get:
      operationId: listEstablishments
      summary: Listar establecimientos del tenant
      security:
        - oauth2: ["organization:read"]
      responses:
        "200":
          description: Establecimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Establishment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createEstablishment
      summary: Crear establecimiento
      security:
        - oauth2: ["organization:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstablishmentInput"
      responses:
        "201":
          description: Establecimiento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Establishment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /emission-points:
    get:
      operationId: listEmissionPoints
      summary: Listar puntos de emision
      security:
        - oauth2: ["organization:read"]
      responses:
        "200":
          description: Puntos de emision
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmissionPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createEmissionPoint
      summary: Crear punto de emision
      security:
        - oauth2: ["organization:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmissionPointInput"
      responses:
        "201":
          description: Punto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmissionPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tax-categories:
    get:
      operationId: listTaxCategories
      summary: Listar categorias tributarias vigentes
      security:
        - oauth2: ["organization:read"]
      responses:
        "200":
          description: Categorias tributarias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxCategory"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tags:
    get:
      operationId: listTags
      summary: Listar etiquetas
      security:
        - oauth2: ["tags:read"]
      responses:
        "200":
          description: Etiquetas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createTag
      summary: Crear etiqueta
      security:
        - oauth2: ["tags:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Etiqueta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /automation/settings:
    get:
      operationId: getAutomationSettings
      summary: Consultar politicas y kill switch
      security:
        - oauth2: ["automation:read"]
      responses:
        "200":
          description: Configuracion efectiva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationSettings"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      operationId: updateAutomationSettings
      summary: Actualizar politicas y kill switch
      security:
        - oauth2: ["automation:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationSettingsInput"
      responses:
        "200":
          description: Configuracion actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationSettings"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /operations/{operationId}:
    get:
      operationId: getOperation
      summary: Consultar estado de operacion asincrona
      security:
        - oauth2: ["operations:read"]
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Estado y resultado normalizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /parties:
    get:
      operationId: searchParties
      summary: Buscar clientes y proveedores
      security:
        - oauth2: ["parties:read"]
      parameters:
        - name: q
          in: query
          schema:
            type: string
            minLength: 2
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/PartyRole"
      responses:
        "200":
          description: Resultados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Party"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createParty
      summary: Crear cliente/proveedor
      security:
        - oauth2: ["parties:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartyInput"
      responses:
        "201":
          description: Contacto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /parties/{partyId}:
    put:
      operationId: updateParty
      summary: Editar cliente/proveedor
      security:
        - oauth2: ["parties:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: partyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartyInput"
      responses:
        "200":
          description: Contacto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /products:
    get:
      operationId: searchProducts
      summary: Buscar productos
      security:
        - oauth2: ["products:read"]
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resultados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createProduct
      summary: Crear producto
      security:
        - oauth2: ["products:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /products/{productId}:
    put:
      operationId: updateProduct
      summary: Editar producto
      security:
        - oauth2: ["products:write"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.invalid/authorize
          tokenUrl: https://auth.example.invalid/token
          scopes:
            context:read: Leer contexto
            parties:read: Leer contactos
            parties:write: Gestionar contactos
            products:read: Leer productos
            products:write: Gestionar productos
            invoices:read: Leer facturas
            invoices:write: Crear y emitir documentos
            invoices:issue: Emitir facturas
            credit-notes:issue: Emitir notas de credito
            receivables:read: Leer cartera
            receivables:write: Registrar cobros y recordatorios
            receivables:notify: Enviar recordatorios
            payables:read: Leer obligaciones
            payables:write: Crear y pagar obligaciones
            payables:extract: Extraer obligaciones desde documentos
            finance:read: Leer resumen financiero
            memberships:read: Leer membresias propias
            service-accounts:read: Leer service accounts
            service-accounts:write: Gestionar service accounts
            organization:read: Leer configuracion de empresa
            organization:write: Gestionar configuracion de empresa
            tags:read: Leer etiquetas
            tags:write: Gestionar etiquetas
            automation:read: Leer politicas de automatizacion
            automation:write: Gestionar politicas y kill switch
            operations:read: Consultar operaciones asincronas
        clientCredentials:
          tokenUrl: https://auth.example.invalid/token
          scopes:
            context:read: Leer contexto
            parties:read: Leer contactos
            parties:write: Gestionar contactos
            products:read: Leer productos
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  schemas:
...
    Membership:
      type: object
      additionalProperties: false
      required: [tenantId, organizationId, ruc, tenantName, roles, active]
      properties:
        tenantId:
          type: string
          format: uuid
        organizationId:
          type: string
        ruc:
          type: string
          pattern: "^[0-9]{13}$"
        tenantName:
          type: string
        roles:
          type: array
          items:
            type: string
        active:
          type: boolean
...
    TenantContext:
      type: object
      additionalProperties: false
      required: [tenantId, ruc, name, roles, scopes, automationWritesEnabled]
      properties:
        tenantId:
          type: string
          format: uuid
        ruc:
          type: string
          pattern: "^[0-9]{13}$"
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        automationWritesEnabled:
          type: boolean
...
  responses:
    Unauthorized:
      description: Token ausente, invalido o sin contexto autorizado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Scope, rol, politica o membresia insuficiente
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer error con scope requerido cuando aplique
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No encontrado o no visible en el tenant
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...